move polyfill code to cpp file
