+++ core/include/vecmem/memory/memory_resource.hpp
@@ -38,54 +38,3 @@ using memory_resource = std::experimental::pmr::memory_resource;
 #else
 #error "C++17 LFTS V1 (P0220R1) component memory_resource not found!?!"
 #endif
-
-#if defined(VECMEM_NOT_HAVE_DEFAULT_RESOURCE)
-
-#include <atomic>
-
-#if defined(VECMEM_HAVE_PMR_MEMORY_RESOURCE)
-namespace std::pmr {
-#else
-namespace std::experimental::pmr {
-#endif
-
-namespace detail {
-class new_delete_resource_impl : public memory_resource {
-    virtual void* do_allocate(std::size_t bytes, std::size_t) override {
-        return malloc(bytes);
-    }
-
-    virtual void do_deallocate(void* p, std::size_t, std::size_t) override {
-        free(p);
-    }
-
-    virtual bool do_is_equal(
-        const memory_resource& other) const noexcept override {
-        return &other == this;
-    }
-};
-
-inline std::atomic<memory_resource*>& default_resource() noexcept {
-    static std::atomic<memory_resource*> res{new_delete_resource()};
-    return res;
-}
-
-}  // namespace detail
-
-inline memory_resource* new_delete_resource() noexcept {
-    static detail::new_delete_resource_impl res{};
-    return &res;
-}
-
-inline memory_resource* get_default_resource() noexcept {
-    return detail::default_resource();
-}
-
-inline memory_resource* set_default_resource(memory_resource* res) noexcept {
-
-    memory_resource* new_res = res == nullptr ? new_delete_resource() : res;
-
-    return detail::default_resource().exchange(new_res);
-}
-}
-#endif
